split tests by api group
