--- packages/evo-sdk-common/tests/workspaces/test_workspace_client.py
@@ -1,777 +1,0 @@
-#  Copyright © 2025 Bentley Systems, Incorporated
-#  Licensed under the Apache License, Version 2.0 (the "License");
-#  you may not use this file except in compliance with the License.
-#  You may obtain a copy of the License at
-#      http://www.apache.org/licenses/LICENSE-2.0
-#  Unless required by applicable law or agreed to in writing, software
-#  distributed under the License is distributed on an "AS IS" BASIS,
-#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#  See the License for the specific language governing permissions and
-#  limitations under the License.
-
-import json
-from unittest import mock
-
-from evo.common import HealthCheckType, RequestMethod, StaticContext
-from evo.common.exceptions import ContextError
-from evo.common.test_tools import MockResponse, TestHTTPHeaderDict, TestWithConnector, utc_datetime
-from evo.common.utils import get_header_metadata
-from evo.workspaces import (
-    AddedInstanceUsers,
-    BulkUserRoleAssignment,
-    BulkUserRoleAssignmentRequest,
-    InstanceUser,
-    OrderByOperatorEnum,
-    User,
-    UserRole,
-    WorkspaceAPIClient,
-    WorkspaceOrderByEnum,
-    WorkspaceRole,
-)
-
-from ..data import load_test_data
-from .consts import (
-    ADMIN_BASE_PATH,
-    BASE_PATH,
-    ORG_UUID,
-    USER_ID,
-)
-from .data import (
-    INSTANCE_ADMIN_ROLE,
-    INSTANCE_USER_1,
-    INSTANCE_USER_2,
-    INSTANCE_USER_3,
-    INSTANCE_USER_ROLE,
-    INVITATION_1,
-    INVITATION_2,
-    TEST_BASIC_WORKSPACE_A,
-    TEST_BASIC_WORKSPACE_B,
-    TEST_BASIC_WORKSPACE_C,
-    TEST_WORKSPACE_A,
-    TEST_WORKSPACE_B,
-    TEST_WORKSPACE_C,
-    TEST_WORKSPACE_D,
-)
-from .helpers import (
-    make_instance_role,
-)
-
-
-class TestWorkspaceClient(TestWithConnector):
-    def setUp(self) -> None:
-        super().setUp()
-        self.workspace_client = WorkspaceAPIClient(connector=self.connector, org_id=ORG_UUID)
-        self.setup_universal_headers(get_header_metadata(WorkspaceAPIClient.__module__))
-
-    def test_from_context(self):
-        client = WorkspaceAPIClient.from_context(StaticContext(connector=self.connector, org_id=ORG_UUID))
-        self.assertIsInstance(client, WorkspaceAPIClient)
-
-    def test_from_context_missing_org_id(self):
-        with self.assertRaises(ContextError):
-            WorkspaceAPIClient.from_context(StaticContext(connector=self.connector))
-
-    async def test_get_service_health(self) -> None:
-        with mock.patch("evo.workspaces.client.get_service_health") as mock_get_service_health:
-            await self.workspace_client.get_service_health()
-        mock_get_service_health.assert_called_once_with(self.connector, "workspace", check_type=HealthCheckType.FULL)
-
-    def _empty_content(self) -> str:
-        data = """{"results": [], "links": {"first": "http://firstlink", "last": "http://lastlink",
-                "next": null, "previous": null, "count": 0, "total": 0}}"""
-        return data
-
-    async def test_list_workspaces_default_args(self):
-        with self.transport.set_http_response(200, self._empty_content(), headers={"Content-Type": "application/json"}):
-            workspaces = await self.workspace_client.list_workspaces()
-        self.assert_request_made(
-            method=RequestMethod.GET,
-            path=f"{BASE_PATH}/workspaces?offset=0",
-            headers={"Accept": "application/json"},
-        )
-        self.assertEqual([], workspaces.items())
-
-    async def test_list_workspaces_all_args(self):
-        with self.transport.set_http_response(200, self._empty_content(), headers={"Content-Type": "application/json"}):
-            workspaces = await self.workspace_client.list_workspaces(offset=10, limit=20)
-        self.assert_request_made(
-            method=RequestMethod.GET,
-            path=f"{BASE_PATH}/workspaces?limit=20&offset=10",
-            headers={"Accept": "application/json"},
-        )
-        self.assertEqual([], workspaces.items())
-
-    async def test_delete_workspace_call(self):
-        with self.transport.set_http_response(204):
-            response = await self.workspace_client.delete_workspace(workspace_id=TEST_WORKSPACE_A.id)
-        self.assert_request_made(method=RequestMethod.DELETE, path=f"{BASE_PATH}/workspaces/{TEST_WORKSPACE_A.id}")
-        self.assertIsNone(response, "Delete workspace response should be None")
-
-    async def test_create_workspace(self):
-        with self.transport.set_http_response(201, json.dumps(load_test_data("new_workspace.json"))):
-            new_workspace = await self.workspace_client.create_workspace(
-                name="Test Workspace",
-                description="test workspace",
-            )
-        self.assert_request_made(
-            method=RequestMethod.POST,
-            path=f"{BASE_PATH}/workspaces",
-            headers={
-                "Accept": "application/json",
-                "Content-Type": "application/json",
-            },
-            body={
-                "bounding_box": None,
-                "default_coordinate_system": "",
-                "description": "test workspace",
-                "labels": None,
-                "name": "Test Workspace",
-            },
-        )
-        self.assertEqual(TEST_WORKSPACE_A, new_workspace)
-
-    async def test_create_workspace_with_bounding_box(self):
-        with self.transport.set_http_response(201, json.dumps(load_test_data("new_workspace_with_bounding_box.json"))):
-            new_workspace = await self.workspace_client.create_workspace(
-                name="Test Workspace",
-                description="test workspace",
-                bounding_box_coordinates=[(100, 0), (101, 0), (101, 1), (100, 1), (100, 0)],
-            )
-        self.assert_request_made(
-            method=RequestMethod.POST,
-            path=f"{BASE_PATH}/workspaces",
-            headers={
-                "Accept": "application/json",
-                "Content-Type": "application/json",
-            },
-            body={
-                "bounding_box": {
-                    "coordinates": [[[100, 0], [101, 0], [101, 1], [100, 1], [100, 0]]],
-                    "type": "Polygon",
-                },
-                "default_coordinate_system": "",
-                "description": "test workspace",
-                "labels": None,
-                "name": "Test Workspace",
-            },
-        )
-        self.assertEqual(TEST_WORKSPACE_D, new_workspace)
-
-    async def test_update_workspace(self):
-        with self.transport.set_http_response(200, json.dumps(load_test_data("new_workspace.json"))):
-            updated_workspace = await self.workspace_client.update_workspace(
-                workspace_id=TEST_WORKSPACE_A.id,
-                name="Test Workspace",
-            )
-        self.assert_request_made(
-            method=RequestMethod.PATCH,
-            path=f"{BASE_PATH}/workspaces/{TEST_WORKSPACE_A.id}",
-            headers={
-                "Accept": "application/json",
-                "Content-Type": "application/json",
-            },
-            body={
-                "name": "Test Workspace",
-            },
-        )
-        self.assertEqual(TEST_WORKSPACE_A, updated_workspace)
-
-    async def test_assign_user_role(self) -> None:
-        with self.transport.set_http_response(
-            201,
-            json.dumps(
-                {
-                    "user_id": str(USER_ID),
-                    "role": "owner",
-                }
-            ),
-        ):
-            response = await self.workspace_client.assign_user_role(
-                workspace_id=TEST_WORKSPACE_A.id,
-                user_id=USER_ID,
-                role=WorkspaceRole.owner,
-            )
-        self.assert_request_made(
-            method=RequestMethod.POST,
-            path=f"{BASE_PATH}/workspaces/{TEST_WORKSPACE_A.id}/users",
-            headers={
-                "Accept": "application/json",
-                "Content-Type": "application/json",
-            },
-            body={
-                "user_id": str(USER_ID),
-                "role": "owner",
-            },
-        )
-        self.assertEqual(response, UserRole(user_id=USER_ID, role=WorkspaceRole.owner))
-
-    async def test_get_current_user_role(self) -> None:
-        with self.transport.set_http_response(
-            200,
-            json.dumps(
-                {
-                    "user_id": str(USER_ID),
-                    "role": "owner",
-                }
-            ),
-        ):
-            response = await self.workspace_client.get_current_user_role(workspace_id=TEST_WORKSPACE_A.id)
-        self.assert_request_made(
-            method=RequestMethod.GET,
-            path=f"{BASE_PATH}/workspaces/{TEST_WORKSPACE_A.id}/current-user-role",
-            headers={"Accept": "application/json"},
-        )
-        self.assertEqual(response, UserRole(user_id=USER_ID, role=WorkspaceRole.owner))
-
-    async def test_list_user_roles(self) -> None:
-        with self.transport.set_http_response(
-            200,
-            json.dumps(
-                {
-                    "results": [
-                        {
-                            "user_id": str(USER_ID),
-                            "role": "owner",
-                            "full_name": "Test User",
-                            "email": "test@example.com",
-                        },
-                    ],
-                    "links": {"self": "dummy-link.com"},
-                }
-            ),
-        ):
-            response = await self.workspace_client.list_user_roles(workspace_id=TEST_WORKSPACE_A.id)
-
-        self.assert_request_made(
-            method=RequestMethod.GET,
-            path=f"{BASE_PATH}/workspaces/{TEST_WORKSPACE_A.id}/users",
-            headers={"Accept": "application/json"},
-        )
-        self.assertEqual(
-            response,
-            [
-                User(user_id=USER_ID, role=WorkspaceRole.owner, full_name="Test User", email="test@example.com"),
-            ],
-        )
-
-    async def test_list_workspaces(self) -> None:
-        content = load_test_data("list_workspaces_0.json")
-        with self.transport.set_http_response(200, json.dumps(content), headers={"Content-Type": "application/json"}):
-            workspaces = await self.workspace_client.list_workspaces(limit=2)
-        self.assert_request_made(
-            method=RequestMethod.GET,
-            path=f"{BASE_PATH}/workspaces?limit=2&offset=0",
-            headers={"Accept": "application/json"},
-        )
-        self.assertEqual(1, self.transport.request.call_count, "One requests should be made.")
-        self.assertEqual([TEST_WORKSPACE_A, TEST_WORKSPACE_B], workspaces.items())
-
-    async def test_list_all_workspaces(self) -> None:
-        content_0 = load_test_data("list_workspaces_0.json")
-        content_1 = load_test_data("list_workspaces_1.json")
-        responses = [
-            MockResponse(status_code=200, content=json.dumps(content_0), headers={"Content-Type": "application/json"}),
-            MockResponse(status_code=200, content=json.dumps(content_1), headers={"Content-Type": "application/json"}),
-        ]
-        self.transport.request.side_effect = responses
-        expected_workspaces = [TEST_WORKSPACE_A, TEST_WORKSPACE_B, TEST_WORKSPACE_C]
-        actual_workspaces = await self.workspace_client.list_all_workspaces(limit=2)
-
-        self.assertEqual(2, self.transport.request.call_count, "Two requests should be made.")
-        self.assert_any_request_made(
-            method=RequestMethod.GET,
-            path=f"{BASE_PATH}/workspaces?limit=2&offset=0",
-            headers=TestHTTPHeaderDict({"Accept": "application/json"}),
-        )
-        self.assert_any_request_made(
-            method=RequestMethod.GET,
-            path=f"{BASE_PATH}/workspaces?limit=2&offset=2",
-            headers=TestHTTPHeaderDict({"Accept": "application/json"}),
-        )
-        self.assertEqual(expected_workspaces, actual_workspaces)
-
-    async def test_list_workspaces_sorted_by_display_name(self) -> None:
-        content_0 = load_test_data("list_workspaces_0.json")
-        content_1 = load_test_data("list_workspaces_1.json")
-
-        # Shuffle the workspaces in the response content.
-        results_0: list = content_0["results"]
-        results_1: list = content_1["results"]
-        results_0.insert(2, results_1.pop(0))
-        results_1.insert(0, results_0.pop(0))
-
-        responses = [
-            MockResponse(status_code=200, content=json.dumps(content_0), headers={"Content-Type": "application/json"}),
-            MockResponse(status_code=200, content=json.dumps(content_1), headers={"Content-Type": "application/json"}),
-        ]
-        self.transport.request.side_effect = responses
-        expected_workspaces = [TEST_WORKSPACE_A, TEST_WORKSPACE_B, TEST_WORKSPACE_C]
-        actual_workspaces = await self.workspace_client.list_all_workspaces(limit=2)
-
-        self.assertEqual(2, self.transport.request.call_count, "Two requests should be made.")
-        self.assert_any_request_made(
-            method=RequestMethod.GET,
-            path=f"{BASE_PATH}/workspaces?limit=2&offset=0",
-            headers=TestHTTPHeaderDict({"Accept": "application/json"}),
-        )
-        self.assert_any_request_made(
-            method=RequestMethod.GET,
-            path=f"{BASE_PATH}/workspaces?limit=2&offset=2",
-            headers=TestHTTPHeaderDict({"Accept": "application/json"}),
-        )
-        self.assertEqual(expected_workspaces, actual_workspaces)
-
-    async def test_list_workspaces_summary_default_args(self):
-        with self.transport.set_http_response(200, self._empty_content(), headers={"Content-Type": "application/json"}):
-            workspaces = await self.workspace_client.list_workspaces_summary()
-        self.assert_request_made(
-            method=RequestMethod.GET,
-            path=f"{BASE_PATH}/workspaces/summary",
-            headers={"Accept": "application/json"},
-        )
-        self.assertEqual([], workspaces.items())
-
-    async def test_list_workspaces_summary_all_args(self) -> None:
-        for order_by in [
-            {WorkspaceOrderByEnum.name: OrderByOperatorEnum.asc},
-            {"name": "asc"},
-        ]:
-            with self.transport.set_http_response(
-                200, self._empty_content(), headers={"Content-Type": "application/json"}
-            ):
-                workspaces = await self.workspace_client.list_workspaces_summary(
-                    offset=10,
-                    limit=20,
-                    order_by=order_by,
-                    filter_created_by=USER_ID,
-                    created_at=str(utc_datetime(2020, 1, 1)),
-                    updated_at=str(utc_datetime(2020, 1, 1)),
-                    name="Test Workspace A",
-                    deleted=False,
-                    filter_user_id=USER_ID,
-                )
-            self.assert_request_made(
-                method=RequestMethod.GET,
-                path=f"{BASE_PATH}/workspaces/summary?"
-                f"limit=20&offset=10&order_by=asc%3Aname&created_by=00000000-0000-0000-0000-000000000002&"
-                f"created_at=2020-01-01+00%3A00%3A00%2B00%3A00&updated_at=2020-01-01+00%3A00%3A00%2B00%3A00&"
-                f"name=Test+Workspace+A&deleted=False&user_id=00000000-0000-0000-0000-000000000002",
-                headers={"Accept": "application/json"},
-            )
-            self.assertEqual([], workspaces.items())
-
-    async def test_list_workspaces_summary(self) -> None:
-        content = load_test_data("list_workspaces_summary.json")
-        with self.transport.set_http_response(200, json.dumps(content), headers={"Content-Type": "application/json"}):
-            workspaces = await self.workspace_client.list_workspaces_summary()
-        self.assert_request_made(
-            method=RequestMethod.GET,
-            path=f"{BASE_PATH}/workspaces/summary",
-            headers={"Accept": "application/json"},
-        )
-        self.assertEqual(1, self.transport.request.call_count, "One requests should be made.")
-        self.assertEqual([TEST_BASIC_WORKSPACE_A, TEST_BASIC_WORKSPACE_B, TEST_BASIC_WORKSPACE_C], workspaces.items())
-
-    async def test_paginated_list_workspaces_summary(self) -> None:
-        content = load_test_data("list_workspaces_summary_paginated_0.json")
-        content_2 = load_test_data("list_workspaces_summary_paginated_1.json")
-        with self.transport.set_http_response(200, json.dumps(content), headers={"Content-Type": "application/json"}):
-            workspaces_page_1 = await self.workspace_client.list_workspaces_summary(limit=2, offset=0)
-
-        with self.transport.set_http_response(200, json.dumps(content_2), headers={"Content-Type": "application/json"}):
-            workspaces_page_2 = await self.workspace_client.list_workspaces_summary(limit=2, offset=2)
-
-        self.assert_any_request_made(
-            method=RequestMethod.GET,
-            path=f"{BASE_PATH}/workspaces/summary?limit=2&offset=0",
-            headers={"Accept": "application/json"},
-        )
-        self.assert_any_request_made(
-            method=RequestMethod.GET,
-            path=f"{BASE_PATH}/workspaces/summary?limit=2&offset=2",
-            headers={"Accept": "application/json"},
-        )
-        self.assertEqual(2, self.transport.request.call_count, "Two requests should be made.")
-        self.assertEqual([TEST_BASIC_WORKSPACE_A, TEST_BASIC_WORKSPACE_B], workspaces_page_1.items())
-        self.assertEqual([TEST_BASIC_WORKSPACE_C], workspaces_page_2.items())
-
-    async def test_list_instance_users(self) -> None:
-        content_1 = load_test_data("instance_users_page_1.json")
-        content_2 = load_test_data("instance_users_page_2.json")
-
-        with self.transport.set_http_response(200, json.dumps(content_1), headers={"Content-Type": "application/json"}):
-            users_page_1 = await self.workspace_client.list_instance_users(limit=2, offset=0)
-
-        with self.transport.set_http_response(200, json.dumps(content_2), headers={"Content-Type": "application/json"}):
-            users_page_2 = await self.workspace_client.list_instance_users(limit=2, offset=2)
-
-        self.assert_any_request_made(
-            method=RequestMethod.GET,
-            path=f"{BASE_PATH}/members?limit=2&offset=0",
-            headers={"Accept": "application/json"},
-        )
-        self.assert_any_request_made(
-            method=RequestMethod.GET,
-            path=f"{BASE_PATH}/members?limit=2&offset=2",
-            headers={"Accept": "application/json"},
-        )
-        self.assertEqual(2, self.transport.request.call_count, "Two requests should be made.")
-        self.assertEqual([INSTANCE_USER_1, INSTANCE_USER_2], users_page_1.items())
-        self.assertEqual([INSTANCE_USER_3], users_page_2.items())
-
-    async def test_list_instance_user_invitations(self) -> None:
-        content = load_test_data("invitations_page_1.json")
-
-        with self.transport.set_http_response(200, json.dumps(content), headers={"Content-Type": "application/json"}):
-            invitations = await self.workspace_client.list_instance_user_invitations(limit=2, offset=0)
-
-        self.assert_request_made(
-            method=RequestMethod.GET,
-            path=f"{BASE_PATH}/members/invitations?limit=2&offset=0",
-            headers={"Accept": "application/json"},
-        )
-        self.assertEqual([INVITATION_1, INVITATION_2], invitations.items())
-
-    async def test_add_users_to_instance(self) -> None:
-        add_users_content = load_test_data("add_instance_users.json")
-
-        with self.transport.set_http_response(
-            201,
-            json.dumps(add_users_content),
-            headers={"Content-Type": "application/json"},
-        ):
-            response = await self.workspace_client.add_users_to_instance(
-                users={
-                    INSTANCE_USER_2.email: [INSTANCE_USER_2.roles[0].role_id],
-                    INVITATION_1.email: [INVITATION_1.roles[0].role_id],
-                }
-            )
-        self.assert_request_made(
-            method=RequestMethod.POST,
-            path=f"{BASE_PATH}/members",
-            headers={
-                "Accept": "application/json",
-                "Content-Type": "application/json",
-            },
-            body={
-                "users": [
-                    {
-                        "email": INSTANCE_USER_2.email,
-                        "roles": [str(INSTANCE_USER_2.roles[0].role_id)],
-                    },
-                    {
-                        "email": INVITATION_1.email,
-                        "roles": [str(INVITATION_1.roles[0].role_id)],
-                    },
-                ]
-            },
-        )
-
-        self.assertEqual(
-            response,
-            AddedInstanceUsers(
-                members=[INSTANCE_USER_2],
-                invitations=[INVITATION_1],
-            ),
-        )
-
-    async def test_delete_instance_user_invitation(self) -> None:
-        with self.transport.set_http_response(204):
-            response = await self.workspace_client.delete_instance_user_invitation(
-                invitation_id=INVITATION_1.invitation_id
-            )
-        self.assert_request_made(
-            method=RequestMethod.DELETE,
-            path=f"{BASE_PATH}/members/invitations/{INVITATION_1.invitation_id}",
-        )
-        self.assertIsNone(response, "Delete instance user invitation response should be None")
-
-    async def test_remove_instance_user(self) -> None:
-        with self.transport.set_http_response(204):
-            response = await self.workspace_client.remove_instance_user(user_id=INSTANCE_USER_1.user_id)
-        self.assert_request_made(
-            method=RequestMethod.DELETE,
-            path=f"{BASE_PATH}/members/{INSTANCE_USER_1.user_id}",
-        )
-        self.assertIsNone(response, "Remove instance user response should be None")
-
-    async def test_update_instance_user_roles(self) -> None:
-        update_users_content = load_test_data("update_instance_user.json")
-        with self.transport.set_http_response(
-            200,
-            json.dumps(update_users_content),
-            headers={"Content-Type": "application/json"},
-        ):
-            response = await self.workspace_client.update_instance_user_roles(
-                user_id=INSTANCE_USER_1.user_id,
-                roles=[INSTANCE_ADMIN_ROLE.role_id],
-            )
-        self.assert_request_made(
-            method=RequestMethod.PATCH,
-            path=f"{BASE_PATH}/members/{INSTANCE_USER_1.user_id}",
-            headers={
-                "Accept": "application/json",
-                "Content-Type": "application/json",
-            },
-            body={
-                "user_id": str(INSTANCE_USER_1.user_id),
-                "roles": [str(INSTANCE_ADMIN_ROLE.role_id)],
-            },
-        )
-        self.assertEqual(
-            response,
-            InstanceUser(
-                user_id=INSTANCE_USER_1.user_id,
-                roles=[make_instance_role(INSTANCE_ADMIN_ROLE.role_id, "Evo Admin")],
-            ),
-        )
-
-    async def test_restore_workspace(self):
-        with self.transport.set_http_response(204):
-            response = await self.workspace_client.restore_deleted_workspace(workspace_id=TEST_WORKSPACE_A.id)
-        self.assert_request_made(
-            method=RequestMethod.POST,
-            path=f"{BASE_PATH}/workspaces/{TEST_WORKSPACE_A.id}?deleted=false",
-            headers={"accept": "application/json"},
-        )
-        self.assertIsNone(response, "Restore workspace response should be None")
-
-    async def test_get_thumbnail(self):
-        thumbnail_bytes: bytearray = load_test_data("thumbnail.jpg")
-        self.transport.request.return_value = MockResponse(
-            status_code=200,
-            body=thumbnail_bytes,
-            headers={"Content-Type": "image/jpeg"},
-        )
-        response = await self.workspace_client.get_thumbnail(workspace_id=TEST_WORKSPACE_A.id)
-        self.assert_request_made(
-            method=RequestMethod.GET,
-            path=f"{BASE_PATH}/workspaces/{TEST_WORKSPACE_A.id}/thumbnail",
-            headers={"accept": "image/jpeg"},
-        )
-        self.assertEqual(response, thumbnail_bytes)
-
-    async def test_put_thumbnail(self):
-        thumbnail_bytes: bytearray = load_test_data("thumbnail.jpg")
-        with self.transport.set_http_response(204):
-            response = await self.workspace_client.put_thumbnail(
-                workspace_id=TEST_WORKSPACE_A.id, thumbnail=thumbnail_bytes
-            )
-        self.assert_request_made(
-            method=RequestMethod.PUT,
-            path=f"{BASE_PATH}/workspaces/{TEST_WORKSPACE_A.id}/thumbnail",
-            headers={"Content-Type": "image/jpeg"},
-            body=thumbnail_bytes,
-        )
-        self.assertIsNone(response, "Put thumbnail response should be None")
-
-    async def test_delete_thumbnail(self):
-        with self.transport.set_http_response(204):
-            response = await self.workspace_client.delete_thumbnail(workspace_id=TEST_WORKSPACE_A.id)
-        self.assert_request_made(
-            method=RequestMethod.DELETE,
-            path=f"{BASE_PATH}/workspaces/{TEST_WORKSPACE_A.id}/thumbnail",
-        )
-        self.assertIsNone(response, "Delete thumbnail response should be None")
-
-    async def test_bulk_assign_roles(self):
-        request = BulkUserRoleAssignmentRequest(
-            role_assignments=[
-                BulkUserRoleAssignment(
-                    workspace_id=TEST_WORKSPACE_A.id,
-                    role=WorkspaceRole.editor,
-                    user_id=INSTANCE_USER_1.user_id,
-                )
-            ]
-        )
-        with self.transport.set_http_response(200):
-            response = await self.workspace_client.admin_bulk_assign_roles(request)
-        self.assert_request_made(
-            method=RequestMethod.POST,
-            path=f"{ADMIN_BASE_PATH}/action/bulk_assign_roles",
-            headers={"Content-Type": "application/json", "Accept": "application/json"},
-            body={
-                "role_assignments": [
-                    {
-                        "role": "editor",
-                        "user_id": str(INSTANCE_USER_1.user_id),
-                        "workspace_id": str(TEST_WORKSPACE_A.id),
-                    }
-                ]
-            },
-        )
-        self.assertIsNone(response, "Bulk assign roles response should be None")
-
-    async def test_admin_list_user_workspaces(self):
-        content = load_test_data("list_user_workspaces_admin.json")
-        with self.transport.set_http_response(200, json.dumps(content), headers={"Content-Type": "application/json"}):
-            workspaces = await self.workspace_client.admin_list_user_workspaces(user_id=INSTANCE_USER_1.user_id)
-        self.assert_request_made(
-            method=RequestMethod.GET,
-            path=f"{ADMIN_BASE_PATH}/users/{INSTANCE_USER_1.user_id}/workspaces?limit=100&offset=0",
-            headers={"Accept": "application/json"},
-        )
-        self.assertEqual(1, self.transport.request.call_count, "One requests should be made.")
-        self.assertEqual([TEST_WORKSPACE_A, TEST_WORKSPACE_B], workspaces.items())
-
-    async def test_admin_list_workspaces(self):
-        content = load_test_data("list_workspaces_0.json")
-        with self.transport.set_http_response(200, json.dumps(content), headers={"Content-Type": "application/json"}):
-            workspaces = await self.workspace_client.admin_list_workspaces()
-            self.assert_request_made(
-                method=RequestMethod.GET,
-                path=f"{ADMIN_BASE_PATH}/workspaces?limit=100&offset=0",
-                headers={"Accept": "application/json"},
-            )
-            self.assertEqual(1, self.transport.request.call_count, "One requests should be made.")
-            self.assertEqual([TEST_WORKSPACE_A, TEST_WORKSPACE_B], workspaces.items())
-
-    async def test_admin_get_workspace(self):
-        content = load_test_data("get_workspace.json")
-        with self.transport.set_http_response(
-            200,
-            json.dumps(content),
-            headers={
-                "Accept": "application/json",
-                "Content-Type": "application/json",
-            },
-        ):
-            workspace = await self.workspace_client.admin_get_workspace(workspace_id=TEST_WORKSPACE_A.id)
-
-        self.assert_request_made(
-            method=RequestMethod.GET,
-            path=f"{ADMIN_BASE_PATH}/workspaces/{TEST_WORKSPACE_A.id}?deleted=False",
-            headers={"Accept": "application/json"},
-        )
-        self.assertEqual(TEST_WORKSPACE_A, workspace)
-
-    async def test_admin_get_thumbnail(self):
-        thumbnail_bytes: bytearray = load_test_data("thumbnail.jpg")
-        self.transport.request.return_value = MockResponse(
-            status_code=200,
-            body=thumbnail_bytes,
-            headers={"Content-Type": "image/jpeg"},
-        )
-        response = await self.workspace_client.admin_get_thumbnail(workspace_id=TEST_WORKSPACE_A.id)
-        self.assert_request_made(
-            method=RequestMethod.GET,
-            path=f"{ADMIN_BASE_PATH}/workspaces/{TEST_WORKSPACE_A.id}/thumbnail",
-            headers={"accept": "image/jpeg"},
-        )
-        self.assertEqual(response, thumbnail_bytes)
-
-    async def test_admin_list_users(self):
-        with self.transport.set_http_response(
-            200,
-            json.dumps(
-                {
-                    "results": [
-                        {
-                            "user_id": str(USER_ID),
-                            "role": "owner",
-                            "full_name": "Test User",
-                            "email": "test@example.com",
-                        },
-                    ],
-                    "links": {"self": "dummy-link.com"},
-                }
-            ),
-        ):
-            response = await self.workspace_client.admin_list_users(workspace_id=TEST_WORKSPACE_A.id)
-        self.assert_request_made(
-            method=RequestMethod.GET,
-            path=f"{ADMIN_BASE_PATH}/workspaces/{TEST_WORKSPACE_A.id}/users",
-            headers={"Accept": "application/json"},
-        )
-        self.assertEqual(
-            response,
-            [
-                User(user_id=USER_ID, role=WorkspaceRole.owner, full_name="Test User", email="test@example.com"),
-            ],
-        )
-
-    async def test_admin_assign_user_role(self):
-        with self.transport.set_http_response(
-            201,
-            json.dumps(
-                {
-                    "user_id": str(USER_ID),
-                    "role": "owner",
-                }
-            ),
-        ):
-            response = await self.workspace_client.admin_assign_user_role(
-                workspace_id=TEST_WORKSPACE_A.id,
-                user_id=USER_ID,
-                role=WorkspaceRole.owner,
-            )
-        self.assert_request_made(
-            method=RequestMethod.POST,
-            path=f"{ADMIN_BASE_PATH}/workspaces/{TEST_WORKSPACE_A.id}/users",
-            headers={
-                "Accept": "application/json",
-                "Content-Type": "application/json",
-            },
-            body={
-                "user_id": str(USER_ID),
-                "role": "owner",
-            },
-        )
-        self.assertEqual(response, UserRole(user_id=USER_ID, role=WorkspaceRole.owner))
-
-    # admin remove user from workspace
-
-    async def test_admin_remove_user_from_workspace(self):
-        with self.transport.set_http_response(204):
-            response = await self.workspace_client.admin_remove_user_from_workspace(TEST_WORKSPACE_A.id, USER_ID)
-
-        self.assert_request_made(
-            method=RequestMethod.DELETE,
-            path=f"{ADMIN_BASE_PATH}/workspaces/{TEST_WORKSPACE_A.id}/users/{USER_ID}",
-        )
-        self.assertIsNone(response, "delete response should be empty.")
-
-    async def test_remove_user_from_workspace(self):
-        with self.transport.set_http_response(204):
-            response = await self.workspace_client.delete_user_role(TEST_WORKSPACE_A.id, USER_ID)
-
-        self.assert_request_made(
-            method=RequestMethod.DELETE,
-            path=f"{BASE_PATH}/workspaces/{TEST_WORKSPACE_A.id}/users/{USER_ID}",
-        )
-        self.assertIsNone(response, "delete response should be empty.")
-
-    async def test_list_instance_user_roles(self):
-        content = load_test_data("list_instance_roles.json")
-        with self.transport.set_http_response(
-            200,
-            json.dumps(content),
-            headers={"Content-Type": "application/json"},
-        ):
-            response = await self.workspace_client.list_instance_roles()
-
-        self.assert_request_made(
-            method=RequestMethod.GET,
-            path=f"{BASE_PATH}/members/roles",
-            headers={"Accept": "application/json"},
-        )
-        self.assertEqual(response, [INSTANCE_USER_ROLE, INSTANCE_ADMIN_ROLE])
-
-    async def test_get_workspace(self):
-        content = load_test_data("get_workspace.json")
-        with self.transport.set_http_response(
-            200,
-            json.dumps(content),
-            headers={
-                "Accept": "application/json",
-                "Content-Type": "application/json",
-            },
-        ):
-            workspace = await self.workspace_client.get_workspace(workspace_id=TEST_WORKSPACE_A.id)
-
-        self.assert_request_made(
-            method=RequestMethod.GET,
-            path=f"{BASE_PATH}/workspaces/{TEST_WORKSPACE_A.id}?deleted=False",
-            headers={"Accept": "application/json"},
-        )
-        self.assertEqual(TEST_WORKSPACE_A, workspace)
